requirements.txt
asgiref==3.2.7
click==7.1.1
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
joblib==0.14.1
nltk==3.5
pandas==1.0.3
pkg-resources==0.0.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.45.0
